+++ indico/types/__init__.py
@@ -1,13 +1,15 @@
-from .base import *
-from .datafile import *
-from .dataset import *
-from .integration import *
-from .jobs import *
-from .model import *
-from .model_export import *
-from .model_group import *
-from .output_file import *
-from .questionnaire import Questionnaire  # noqa
-from .submission import *
-from .submission_file import *
-from .workflow import *
+# the order of these imports is super important for preventing import cycles
+
+from .base import *  # noqa
+from .datafile import *  # noqa
+from .dataset import *  # noqa
+from .integration import *  # noqa
+from .jobs import *  # noqa
+from .model import *  # noqa
+from .model_group import *  # noqa
+from .model_export import *  # noqa
+from .output_file import *  # noqa
+from .questionnaire import Example  # noqa
+from .submission import *  # noqa
+from .submission_file import *  # noqa
+from .workflow import *  # noqa
