+++ tox.ini
@@ -14,6 +14,6 @@ deps =
     typing_extensions
 commands =
     # NOTE: you can run any command line tool here - not just testsg
-    mypy --install-types
+    mypy --install-types --non-interactive
     mypy --config-file=pyproject.toml
     pytest -sv tests/unit
